+++ History.md
@@ -2,6 +2,7 @@
 ==================
 
 * Drops support for Node.js 6 & 8 to support last 3 Node.js releases: 10, 12, and 14.
+* Fixed issue [#889](https://github.com/jakubpawlowicz/clean-css/issues/889) - whitelisted level 1 optimizations.
 
 [4.2.2 / 2020-01-25](https://github.com/jakubpawlowicz/clean-css/compare/v4.2.1...v4.2.2)
 ==================